third_party/blink/web_tests/http/tests/websocket/empty-ping-expected.txt

Test whether WebSocket correctly responds to an empty ping message sent from the server.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onmessage() was called. (message = "PASS")
PASS message is "PASS"
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
